Process valid lines where possible
The current implementation will drop all points in a batch if any of them are invalid.  Instead, process all the valid lines and drop only the invalid ones.  Keep the existing behavior of returning StatusBadRequest if any lines are invalid and return a json response with success / failure counts for the batch.